Have fixed thing

+++ eventActions/backspeak.js
@@ -10,14 +10,16 @@ class backspeakCheckAction {
 				gameName: "backspeak"
 			}
 		}).then(result => {
-			var validResp = result[0].content;
-			if (message.content == validResp) {
-				message.channel.send("And the winner is... " + message.author + "!");
-				Backspeak.destroy({
-					where: {
-						gameName: "backspeak"
-					}
-				});
+			if (result.length > 0) {
+				var validResp = result[0].content;
+				if (message.content == validResp) {
+					message.channel.send("And the winner is... " + message.author + "!");
+					Backspeak.destroy({
+						where: {
+							gameName: "backspeak"
+						}
+					});
+				}
 			}
 		}));
   }
